refactor: split ui/ into helpers and tui model subpackages
Move bubbletea Model types (SelectorModel, SnapshotEditorModel,
ConfigCustomizerModel, MacOSSelectorModel) and their views/tests to
internal/ui/tui/. The parent ui/ package keeps output helpers,
huh form wrappers, StickyProgress, ScanProgress, and ScrollRegion.

Four callers updated to import the new tui subpackage:
- internal/installer/plan.go (RunSelector, RunMacOSSelector)
- internal/cli/install.go (RunConfigCustomizer)
- internal/cli/snapshot.go (RunSnapshotEditor)
- internal/cli/snapshot_import.go (RunSnapshotEditor)

Archtest baselines regenerated for shifted line numbers; no new
violations introduced.
