+++ internal/archtest/baseline/no-raw-http.txt
@@ -3,7 +3,7 @@
 # Regenerate: ARCHTEST_UPDATE_BASELINE=1 go test ./internal/archtest/...
 internal/auth/login.go:106
 internal/auth/login.go:164
-internal/cli/snapshot_import.go:70
+internal/cli/snapshot_import.go:71
 internal/cli/snapshot_publish.go:142
 internal/config/packages_remote.go:82
 internal/config/packages_remote.go:87
