+++ internal/ui/AGENTS.md
@@ -1,34 +1,33 @@
 # UI PACKAGE
 
-TUI components built on Charmbracelet (bubbletea + lipgloss + huh). 5 files, 2,206 lines.
+Output helpers, progress widgets, and form wrappers. Bubbletea TUI models
+live in the `tui/` subpackage.
 
-## FILES
+## FILES (internal/ui/)
 
 | File | Lines | Purpose |
 |------|-------|---------|
-| `selector.go` | 986 | Package selector: tabs, fuzzy search, online search, multi-select |
-| `snapshot_editor.go` | 518 | Snapshot editing: diff view, toggle packages, confirm |
-| `progress.go` | 289 | StickyProgress: per-package timing, succeeded/failed/skipped counters |
-| `scanprogress.go` | 228 | Scan progress: step timing, overall counter `[3/8]` |
-| `ui.go` | 185 | Base styles, form helpers (SelectPreset, InputGitConfig, Confirm) |
+| `ui.go` | 247 | Base styles, color helpers, output helpers (Header/Success/Error/Info/Warn/Muted/DryRun*), huh form wrappers (InputGitConfig, SelectPreset, Confirm, SelectOption, Input) |
+| `progress.go` | 246 | StickyProgress: per-package timing, succeeded/failed/skipped counters |
+| `scanprogress.go` | 221 | ScanProgress: step timing, overall counter `[3/8]` |
+| `scrollregion.go` | 143 | ANSI scroll-region plumbing for StickyProgress |
+
+## FILES (internal/ui/tui/)
+
+| File | Lines | Purpose |
+|------|-------|---------|
+| `selector.go` + `selector_view.go` | ~1,026 | Package selector: tabs, fuzzy search, online search, multi-select |
+| `snapshot_editor.go` + `snapshot_editor_search.go` | ~1,022 | Snapshot editing: diff view, toggle packages, add online, confirm |
+| `config_customizer.go` | 253 | Remote-config customizer: toggle which packages to install |
+| `macos_selector.go` | 435 | macOS preferences selector: category tabs, toggle, confirm |
 
 ## PATTERNS
 
-- **bubbletea Model**: `selector.go` and `snapshot_editor.go` implement `tea.Model` (Init/Update/View)
+- **bubbletea Model**: `tui/*.go` files implement `tea.Model` (Init/Update/View)
 - **Sticky output**: `progress.go` writes directly to stderr via ANSI escape codes, not bubbletea
 - **Styles**: All lipgloss styles defined as package-level vars at top of file
 - **Color palette**: Primary=#22c55e, Subtle=#666, Warning=#eab308, Danger=#ef4444
 - **Width adaptation**: Components read `tea.WindowSizeMsg` and adapt layout
-
-## SELECTOR ARCHITECTURE (selector.go)
-
-Two render modes: normal (tab navigation) and search (`/` key).
-
-- **Tab bar**: Sliding window — shows active tab + neighbors + `N/M` position. Adapts to terminal width.
-- **Search**: Fuzzy local match (sahilm/fuzzy) + debounced online search (500ms). Animated spinner during fetch.
-- **Toast**: 1.5s auto-fade notifications on toggle (`+ Added node` / `- Removed node`).
-- **Scroll**: `scrollOffset` + `getVisibleItems()` for height-adaptive list (5-20 items).
-- **Alt screen**: Uses `tea.WithAltScreen()` for full-screen mode.
 
 ## PROGRESS ARCHITECTURE (progress.go)
 
@@ -41,7 +40,7 @@ NOT a bubbletea model. Direct stderr writer for use during brew install.
 
 ## WHEN ADDING NEW UI
 
-1. Interactive full-screen → bubbletea Model in new file
-2. Progress/status during install → extend StickyProgress or ScanProgress
-3. Simple form input → use helpers in ui.go (huh-based)
-4. Styled text output → use ui.Header/Success/Error/Info/Warn/Muted
+1. Interactive full-screen → bubbletea Model in `tui/` subpackage
+2. Progress/status during install → extend StickyProgress or ScanProgress in `ui/`
+3. Simple form input → use helpers in `ui.go` (huh-based)
+4. Styled text output → use `ui.Header`/`Success`/`Error`/`Info`/`Warn`/`Muted`
