+++ internal/ui/tui/config_customizer.go
@@ -1,4 +1,4 @@
-package ui
+package tui
 
 import (
 	"fmt"
