+++ internal/ui/tui/config_customizer_test.go
@@ -1,4 +1,4 @@
-package ui
+package tui
 
 import (
 	"testing"
