+++ internal/ui/tui/macos_selector.go
@@ -1,4 +1,4 @@
-package ui
+package tui
 
 import (
 	"fmt"
