+++ internal/ui/tui/macos_selector_test.go
@@ -1,4 +1,4 @@
-package ui
+package tui
 
 import (
 	"strings"
