+++ internal/ui/tui/snapshot_editor.go
@@ -1,4 +1,4 @@
-package ui
+package tui
 
 import (
 	"fmt"
