+++ internal/ui/tui/snapshot_editor_test.go
@@ -1,4 +1,4 @@
-package ui
+package tui
 
 import (
 	"fmt"
